fix: PreAlarmState now works correctly

+++ include/DroneInsideState.h
@@ -18,7 +18,7 @@ class DroneInsideState : public GenericState
         NewPing &sonarUsed,
         int pirState,
         uint8_t analog_pin,
-        int beta
+        float beta
     );
 
     ~DroneInsideState();
